ADDED Cap'n Proto support.

+++ services/PixyzReader/src/batch/vector3_array.py
@@ -3,9 +3,9 @@ class Vector3Array:
         self.vector3_array = []
         for item in array:
             VectorData = {}
-            VectorData["x"] = item.x / divider
-            VectorData["y"] = item.y / divider
-            VectorData["z"] = item.z / divider
+            VectorData["x"] = float(item.x / divider)
+            VectorData["y"] = float(item.y / divider)
+            VectorData["z"] = float(item.z / divider)
             self.vector3_array.append(VectorData)
 
     def Get(self):
